changing seeder files
