+++ server/seeders/main.js
@@ -72,7 +72,7 @@ function generateData() {
         email: email,
         password: password,
         phone: phone,
-        services: services,
+       // services: services,
       });
     }
     console.log(users);
